Make server be threaded
This makes the server resilient to getting an empty TCP request
and getting stuck waiting for the "METHOD /path HTTP/1.x\r\n"
line. Which it maybe seems like IE 11 was doing sometimes,
unpredictably, or maybe IE was getting the server stuck for
some other reason.

Somehow, making it threaded has not led to repeated empty(?) requests
leaving many stuck threads around.

(This resilience is obviously not resilience to actual attackers
who want to DOS it -- to stop that, either don't listen on ports
they can talk to, or use an actual server like nginx that
has code to deal with such things.)

This also means the server has to be equipped to handle requests
running in parallel and/or receiving new POST requests out of order.