fix textarea width
Apparently cols= doesn't compute as precisely as 'ch' unit
does.  In Firefox the cols=80 monospace textarea was 81 or 82
characters wide depending on zoom.  I'm guessing that "cols"
takes the font width which is in non-integer pixels, rounds it
to an integer pixel width, and multiplies it by the value of
cols, or something along those lines, while 'ch' does less
rounding somehow (tested in Chrome and Firefox so far).

'ch' is compatible with modern browsers, IE >= 9
https://developer.mozilla.org/en-US/docs/Web/CSS/length