improve performance on somewhat large files
Tested on a 1000 line 300 kB SVG in Chromium

Spellcheck isn't a great loss for code, though it could be
for other things.

Things I changed are based on hints from Firefox and Chromium
bugs about textareas being slow, that I thought might help.

Just clicking to select another place in the textarea is slow,
so it's not going to be a great situation regardless. I added
a maximum editable file size.