Add ability to reload the page
while keeping state and not talking to the server (aside from
possibly reloading the HTML itself)

For testing purposes, allow this to be triggered by clicking on
the notes in the left bar.

Also, refactor/re-order some of the effects after loading
text into the textarea.  (re-order to be in an order that
seems to potentially make more sense to me although I don't
know whether browsers care.)