Give up on focusing using "autofocus" in favor of Javascript :-/
On Firefox, the display:none prevented autofocus from
autofocusing an element even when the display:none was removed
before DOMReady.   Chrome was fine.

On both Firefox and Chrome, removing one element with autofocus
and adding autofocus to another element, all before DOMReady,
did not lead to the latter element being focused.

So, not worth it for now with two possible elements that I could
end up wanting to focus.  If there ends up being a specific benefit
of autofocus to users, that I can test, then I'll look into
bringing it back.