For Dragon, we have to reload the page after any textarea resize
so don't resize it as often, to reduce the amount of page reloads

editor.scrollHeight + (line_height * 200) + 1:
200 lines is arbitrary
Chrome was more consistent if I added the one pixel

debounce 40ms: 40 is kind of arbitrary. Intentionally less than
the compute line number's 50ms because computing line numbers isn't
useful if we are going to reload.  But not critical if
they happen in the other order.