Don't reload page for dragon if using Web Speech API
This is not a perfect estimate but it will do for now.
For the most part, the Web Speech API does not work
in the same browsers as Dragon does:

The Dragon plugins for Chrome and Firefox do not work
as well as the IE plugin.  IE does not implement
the Web Speech API. (IE 11 is the last version of IE.
Edge doesn't support Web Speech yet either, but even if
it did, Nuance doesn't have Full Text Control for
Edge yet because Edge doesn't offer the same plugin API
that IE does).