Try to make it so you don't edit the DOM when nothing is selected
It could use some more refinement:
what if a checkbox is selected? not editable plz
but on the other hand:
what if a contentEditable area is selected?
(what's the best way to detect that?)
And should we send synthetic keydown/up events, too?
That might be helpful for key-navigated sites, even if
we're pointed at an un-editable part of the DOM, and
miiiight make it work with fancy editor things like CodeMirror.