bililiteRange.util: make findBack more correct for greedy regexps
Previously, (pretend "|" is range position), "aaaa|aa" findBack(/a+/)
would not be able to find anything because the regexp would
always go past the current location by mistake.  Luckily there
is a workaround.  /a+(?=.{2})/ says only to find matches followed
by at least two characters (two being the number of remaining
characters after the current range).  The full workaround'ed regex
is /(?:a+)(?=[\\s\\S]{2})/g, because:
    - /./ can't match newlines in JS,
    - the original pattern might have /|/ that we need to
        parenthesize in order to add a suffix to the full
        original pattern.

Meanwhile, this change allows to simplify the forward-search
code as well.  This simplification makes findprimitive
also work with greedy regex for searches on a range whose
bounds[1] < text.length (relevant if we ever expose that capability).